requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
mistralai
pydantic>=2
pytest

